requirements.txt
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.19.0
opencv-python>=4.5.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-multipart>=0.0.5
psutil>=5.8.0
pyyaml>=5.4.0
pytest>=6.2.0
requests>=2.25.0
pillow>=8.0.0
